requirements.txt
flask
joblib
scikit-learn
google-generativeai
openai